+++ lib/VitableConnect/employers/types/get_invoice_pdf_employers_request.rb
@@ -7,6 +7,8 @@ module VitableConnect
         field :employer_id, -> { String }, optional: false, nullable: false
 
         field :invoice_id, -> { String }, optional: false, nullable: false
+
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
       end
     end
   end