+++ lib/VitableConnect/employers/types/list_payroll_deduction_statements_employers_request.rb
@@ -9,6 +9,8 @@ module VitableConnect
         field :limit, -> { Integer }, optional: true, nullable: false
 
         field :page, -> { Integer }, optional: true, nullable: false
+
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
       end
     end
   end