+++ lib/VitableConnect/employers/types/update_employer_settings_request.rb
@@ -6,6 +6,8 @@ module VitableConnect
       class UpdateEmployerSettingsRequest < Internal::Types::Model
         field :employer_id, -> { String }, optional: false, nullable: false
 
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
+
         field :pay_frequency, -> { VitableConnect::Types::DeductionFrequency }, optional: false, nullable: false
       end
     end