+++ lib/VitableConnect/enrollments/types/get_enrollments_request.rb
@@ -5,6 +5,8 @@ module VitableConnect
     module Types
       class GetEnrollmentsRequest < Internal::Types::Model
         field :enrollment_id, -> { String }, optional: false, nullable: false
+
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
       end
     end
   end