+++ lib/VitableConnect/enrollments/types/terminate_enrollment_request.rb
@@ -6,6 +6,8 @@ module VitableConnect
       class TerminateEnrollmentRequest < Internal::Types::Model
         field :enrollment_id, -> { String }, optional: false, nullable: false
 
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
+
         field :reason, -> { String }, optional: true, nullable: false
 
         field :ticket_number, -> { String }, optional: true, nullable: false