+++ lib/VitableConnect/members/types/list_id_cards_members_request.rb
@@ -5,6 +5,8 @@ module VitableConnect
     module Types
       class ListIDCardsMembersRequest < Internal::Types::Model
         field :member_id, -> { String }, optional: false, nullable: false
+
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
       end
     end
   end