+++ lib/VitableConnect/members/types/list_members_request.rb
@@ -9,6 +9,8 @@ module VitableConnect
         field :page, -> { Integer }, optional: true, nullable: false
 
         field :search, -> { String }, optional: true, nullable: false
+
+        field :vitable_organization, -> { String }, optional: true, nullable: false, api_name: "X-Vitable-Organization"
       end
     end
   end